Corrected Magstripe project to work correctly on the latest hardware.
Fixed BCERRI not being read correctly in all situations - switched to interrupt-driven checks at all times.

Fixed AudioOut demo to use the correct Fast PWM mode.

git-svn-id: http://lufa-lib.googlecode.com/svn@403 d5102386-fcda-11dd-9fdb-3debd5008f28

+++ Demos/Joystick/Joystick.c
@@ -196,7 +196,7 @@ bool GetNextReport(USB_JoystickReport_Data_t* ReportData)
 	  ReportData->Button |= (1 << 0);
 	  
 	/* Check if the new report is different to the previous report */
-	InputChanged = PrevJoyStatus ^ JoyStatus_LCL;
+	InputChanged = (uint8_t)(PrevJoyStatus ^ JoyStatus_LCL);
 
 	/* Save the current joystick status for later comparison */
 	PrevJoyStatus = JoyStatus_LCL;
