+++ LUFA/Drivers/Board/STK525/Joystick.h
@@ -92,7 +92,7 @@
 			static inline uint8_t Joystick_GetStatus(void) ATTR_WARN_UNUSED_RESULT;
 			static inline uint8_t Joystick_GetStatus(void)
 			{
-				return ((~PINB & JOY_BMASK) | ((~PINE & JOY_EMASK) >> 1));
+				return (((uint8_t)~PINB & JOY_BMASK) | (((uint8_t)~PINE & JOY_EMASK) >> 1));
 			}
 		#endif
